services/edge-nodes/python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "r3l-edge"
version = "0.1.0"
description = "R3L Edge Node CLI — verify files locally, attest on-chain"
requires-python = ">=3.10"
dependencies = [
    "pynacl>=1.5.0",
    "base58>=2.1.0",
    "py-tlsh>=4.7.2",
    "open-clip-torch>=2.24.0",
    "torch>=2.0",
    "torchvision>=0.15",
    "timm>=0.9",
    "pillow>=9.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/r3l_edge"]

[tool.hatch.build]
artifacts = ["src/r3l_edge/bin/*"]

[project.scripts]
r3l-edge = "r3l_edge.cli:main"
